refactor: absorb fleet-style into fleet-tui/style and retire the standalone package
The fleet-agnostic framing on a private monorepo TUI package was over-
engineered for the actual scope, so fleet-tui now owns the Fleet brand
style assets under a new ./style subpath and the separate fleet-style
workspace package is dropped along with its overhead.
