--- packages/fleet-style/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@dotobokuri/fleet-style",
-  "version": "1.0.0",
-  "private": true,
-  "type": "module",
-  "files": ["dist", "AGENTS.md", "README.md"],
-  "scripts": {
-    "build": "tsc -p tsconfig.json"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "devDependencies": {
-    "@types/node": "^24.10.0",
-    "typescript": "^6.0.2"
-  }
-}
