+++ packages/fleet-tui/src/style/banner.ts
