+++ packages/fleet-tui/src/style/palette.ts
