+++ runtime/fleet-cli/package.json
@@ -38,7 +38,6 @@
     "@dotobokuri/fleet-carriers": "workspace:*",
     "@dotobokuri/fleet-infra": "workspace:*",
     "@dotobokuri/fleet-mcp-server": "workspace:*",
-    "@dotobokuri/fleet-style": "workspace:*",
     "@dotobokuri/fleet-tui": "workspace:*",
     "@dotobokuri/fleet-wiki": "workspace:*",
     "@dotobokuri/fleet-wiki-ui": "workspace:*",
