+++ runtime/fleet-wiki-ui/package.json
@@ -44,7 +44,7 @@
     "typecheck": "tsc --noEmit -p tsconfig.json && tsc --noEmit -p client/tsconfig.json"
   },
   "dependencies": {
-    "@dotobokuri/fleet-style": "workspace:*",
+    "@dotobokuri/fleet-tui": "workspace:*",
     "@dotobokuri/fleet-wiki": "workspace:*",
     "dompurify": "^3.4.2",
     "highlight.js": "^11.11.1",
